Reestructure `rm` command
By suggestion of @Diegovsky (thanks!), I changed the rm implementation
to use the concept of holes. This means that, instead of removing the
selected tasks from the list data structure, which may mess up the
indexes of the remaining tasks pretty badly, it just sets the their
states as TASK_HOLE. Tasks with this state are not saved to the TODO
file. This hole-based approach is far more robust than the original. It
makes for more elegant code and effortlessly eliminates a whole class
of rather nasty bugs.