Refactor main by introducing handlers
I've introduced the concept of a handler, which is a function that
carries out a command. I've defined one for each existing command. All
handlers have a uniform signature, which involves returning a
Handler_res. This value communicates to the main function what must be
done after the handler is called. It can represent an error, in which
case the exit code is set accordingly.