Add `uncheck` command
I'd say it's worth writing down what the process to add a new command
looks like right now. Here's how it goes:

- Add a new CM_* constant to the Command enum, in parse.h.
- Add a new branch to the check_command function with the name of the
  new command, in parse.c.
- Add a handler for the new command in handler.h and implement it in
  handler.c.
- Finally, add a new branch to the command switch-case in main.c,
  dealing with the new CM_* constant by calling the new handler.

It's possible to eliminate the last step by putting all of the handlers
in an array, indexed by the CM_* constants they correspond to. Another
improvement that could be made would be to remove the code repetition in
handler.c. The problem is that several commands expect their arguments
to be indexes, and they have to do exactly the same thing to parse and
validate them.

I'd tackle both refactorizations in the following commits.