Add `work-on` command
This command marks tasks as being a work-in-progress. There may only be
one work-in-progress task at any given moment; if there is already
another wip task in the list, this command marks it as done.

Now that I've done the refactors I mentioned some commits ago, here's
the step-by-step process to add new commands:

- Add a new CM_* constant to the Command enum, in commands.h.
- Add an entry for this constant to the commands array, in commands.c.
- Add a new handler to handler.h and implement it in handler.c.
- Add the new handler to the handler array, in main.c.
- Update the usage text, in help.c.

Conclusion: the refactors paid off