Final commit
There's a lot more I could do on this project, but honestly, I feel like
I've already gotten what I wanted from it. The program is small, but it
does what it promises, and does it well. My mission is complete, and now
I feel the urge to move on to other, more elaborate projects. It was
pretty fun while it lasted though. Farewell!